+++ packages/components-extra/src/components/Newsletter/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './Newsletter'
+export * from './Newsletter'